moves global context providers into own component and includes toaster
