+++ client/components/pages/submission-system/layout.js
@@ -20,7 +20,7 @@ import Modal from 'uikit/Modal';
 import ToastStack from 'uikit/notifications/ToastStack';
 import { TOAST_VARIANTS } from 'uikit/notifications/Toast';
 import { NOTIFICATION_INTERACTION_EVENTS } from 'uikit/notifications/Notification';
-import { ToasterContext, useToaster, useToastState } from './toaster';
+import { ToasterContext, useToaster, useToastState } from '../../toaster';
 
 const modalPortalRef = React.createRef();
 const useMounted = () => {
@@ -63,7 +63,7 @@ const SubmissionLayout = ({
   children?: React.Element<any>,
   subtitle?: string,
 }) => {
-  const toaster = useToastState();
+  const toaster = useToaster();
   return (
     <ToasterContext.Provider value={toaster}>
       <PageContainer>
