+++ client/pages/_app.js
@@ -3,24 +3,14 @@ import * as React from 'react';
 import nextCookies from 'next-cookies';
 import Router from 'next/router';
 import Cookies from 'js-cookie';
-import Link from 'next/link';
-import { ApolloProvider } from 'react-apollo';
-import { createHttpLink } from 'apollo-link-http';
-import ApolloClient from 'apollo-client';
-import urlJoin from 'url-join';
-import { ApolloProvider as ApolloHooksProvider } from 'react-apollo-hooks';
 import fetch from 'isomorphic-fetch';
 
-import Button from 'uikit/Button';
-import { ThemeProvider } from 'uikit';
 import { EGO_JWT_KEY } from 'global/constants';
 import { LOGIN_PAGE_PATH } from 'global/constants/pages';
-import { NODE_ENV, ENVIRONMENTS, GATEWAY_API_ROOT, AUTH_DISABLED } from 'global/config';
-import { isValidJwt, decodeToken } from 'global/utils/egoJwt';
+import { AUTH_DISABLED } from 'global/config';
+import { isValidJwt } from 'global/utils/egoJwt';
 import getApolloCacheForQueries from 'global/utils/getApolloCacheForQueries';
-import createInMemoryCache from 'global/utils/createInMemoryCache';
 
-import Error401Page from 'components/pages/401';
 import type {
   PageWithConfig,
   GetInitialPropsContext,
@@ -28,8 +18,8 @@ import type {
 } from 'global/utils/pages/types';
 
 import { ERROR_STATUS_KEY } from './_error';
-import { PageContext } from 'global/hooks/usePageContext';
 import useAuthContext from 'global/hooks/useAuthContext';
+import GlobalContextProviders from 'components/GlobalContextProviders';
 
 const enforceLogin = ({ ctx }: { ctx: GetInitialPropsContext }) => {
   const loginRedirect = `${LOGIN_PAGE_PATH}?redirect=${encodeURI(ctx.asPath)}`;
@@ -63,18 +53,6 @@ const Root = (
     }
   });
 
-  const client = new ApolloClient({
-    // $FlowFixMe apollo-client and apollo-link-http have a type conflict in their typing
-    link: createHttpLink({
-      uri: urlJoin(GATEWAY_API_ROOT, '/graphql'),
-      fetch: fetch,
-      headers: {
-        authorization: `Bearer ${egoJwt}`,
-      },
-    }),
-    cache: createInMemoryCache().restore(apolloCache),
-  });
-
   return (
     <>
       <style>
@@ -96,15 +74,9 @@ const Root = (
             }
         `}
       </style>
-      <PageContext.Provider value={ctx}>
-        <ApolloProvider client={client}>
-          <ApolloHooksProvider client={client}>
-            <ThemeProvider>
-              <Component egoJwt={egoJwt} logOut={logOut} pathname={pathname} {...pageProps} />
-            </ThemeProvider>
-          </ApolloHooksProvider>
-        </ApolloProvider>
-      </PageContext.Provider>
+      <GlobalContextProviders egoJwt={egoJwt} apolloCache={apolloCache} pageContext={ctx}>
+        <Component egoJwt={egoJwt} logOut={logOut} pathname={pathname} {...pageProps} />
+      </GlobalContextProviders>
     </>
   );
 };
